third_party/blink/web_tests/external/wpt/css/css-inline/baseline-source/baseline-source-last-textarea-003.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] .target > * 1
  assert_equals: \n<span data-offset-x="45"></span>\noffsetLeft expected 45 but got 30
[FAIL] .target > * 3
  assert_equals: \n<span data-offset-x="45"></span>\noffsetLeft expected 45 but got 90
[FAIL] .target > * 5
  assert_equals: \n<span data-offset-x="45"></span>\noffsetLeft expected 45 but got 30
[FAIL] .target > * 7
  assert_equals: \n<span data-offset-x="45"></span>\noffsetLeft expected 45 but got 65
[FAIL] .target > * 9
  assert_equals: \n<span data-offset-x="45"></span>\noffsetLeft expected 45 but got 90
[FAIL] .target > * 11
  assert_equals: \n<span data-offset-x="45"></span>\noffsetLeft expected 45 but got 65
Harness: the test ran to completion.
